fix time type

+++ src/VCDTypes.hpp
@@ -22,7 +22,7 @@ typedef std::string VCDScopeName;
 typedef std::string VCDSignalHash;
 
 //! Represents a single instant in time in a trace
-typedef double VCDTime;
+typedef long long VCDTime;
 
 //! Specifies the timing resoloution along with VCDTimeUnit
 typedef unsigned VCDTimeRes;
